move computeBirectionalTraces
